レビューを受けて修正

+++ app/controllers/items_controller.rb
@@ -24,8 +24,12 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if current_user.id == item.user_id
+      item.destroy
+      redirect_to root_path
+    else
+      redirect_to new_user_registration_path
+    end
   end
 
   private
